Revert "Scratch/user authentication (#5)"
This reverts commit 2c9d19c2bfe661fdc8a50a1c643552b859d02509.
