+++ .npmrc
@@ -1,1 +1,2 @@
 package-lock=true
+package-lock=true
