--- amplify/backend/auth/ZeroCarbon/assets/index.html
@@ -1,24 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="UTF-8">
-    <title>AWS SDK for JavaScript - Browser Getting Started Application</title>
-    <link rel="stylesheet" type="text/css" href="style.css">
-
-  </head>
- 
-
-  <body>
-    <div>
-      <div id="myspinner"></div>
-    </div>
-    <script src="https://sdk.amazonaws.com/js/aws-sdk-2.474.0.min.js"></script>
-    <script src="./spinner.js" type="text/javascript"></script>
-    <script src="./verify.js" type="text/javascript"></script>
-    <script type="text/javascript">
-        window.onload = function() {
-            confirm();
-        };
-    </script>
-  </body>
-</html>
