--- amplify/backend/auth/ZeroCarbon/parameters.json
@@ -1,76 +1,0 @@
-{
-    "identityPoolName": "ZeroCarbon",
-    "allowUnauthenticatedIdentities": false,
-    "resourceNameTruncated": "ZeroCab6dfaf0c",
-    "userPoolName": "ZeroCarbon",
-    "autoVerifiedAttributes": [
-        "email"
-    ],
-    "mfaConfiguration": "OFF",
-    "mfaTypes": [
-        "SMS Text Message"
-    ],
-    "smsAuthenticationMessage": "Your authentication code is {####}",
-    "smsVerificationMessage": "Your verification code is {####}",
-    "emailVerificationSubject": "ZeroCarbon: Verification Code",
-    "emailVerificationMessage": "Your verification code is {####}",
-    "defaultPasswordPolicy": false,
-    "passwordPolicyMinLength": 8,
-    "passwordPolicyCharacters": [],
-    "requiredAttributes": [
-        "email",
-        "name"
-    ],
-    "userpoolClientGenerateSecret": true,
-    "userpoolClientRefreshTokenValidity": "365",
-    "userpoolClientWriteAttributes": [
-        "email",
-        "name"
-    ],
-    "userpoolClientReadAttributes": [
-        "email",
-        "name"
-    ],
-    "userpoolClientLambdaRole": "ZeroCab6dfaf0c_userpoolclient_lambda_role",
-    "userpoolClientSetAttributes": true,
-    "useDefault": "manual",
-    "authSelections": "identityPoolAndUserPool",
-    "resourceName": "ZeroCarbon",
-    "thirdPartyAuth": false,
-    "usernameAttributes": [
-        "email"
-    ],
-    "userPoolGroups": false,
-    "adminQueries": false,
-    "triggers": "{\"CustomMessage\":[\"verification-link\"]}",
-    "hostedUI": false,
-    "userPoolGroupList": [],
-    "verificationBucketName": "zerocarbonverificationbucket",
-    "authRoleArn": {
-        "Fn::GetAtt": [
-            "AuthRole",
-            "Arn"
-        ]
-    },
-    "unauthRoleArn": {
-        "Fn::GetAtt": [
-            "UnauthRole",
-            "Arn"
-        ]
-    },
-    "parentStack": {
-        "Ref": "AWS::StackId"
-    },
-    "permissions": [],
-    "dependsOn": [
-        {
-            "category": "function",
-            "resourceName": "ZeroCarbonCustomMessage",
-            "triggerProvider": "Cognito",
-            "attributes": [
-                "Arn",
-                "Name"
-            ]
-        }
-    ]
-}
