+++ amplify/backend/auth/institutoterrabc9982ab/institutoterrabc9982ab-cloudformation-template.yml
@@ -9,18 +9,6 @@ Parameters:
     Type: String
   
   
-  
-  
-  functionZeroCarbonCustomMessageArn: 
-    Type: String
-    Default: functionZeroCarbonCustomMessageArn
-  
-  functionZeroCarbonCustomMessageName: 
-    Type: String
-    Default: functionZeroCarbonCustomMessageName
-  
-  
-  
 
     
   identityPoolName:
@@ -70,10 +58,10 @@ Parameters:
         
   userpoolClientGenerateSecret:
     Type: String
-            
+                
   userpoolClientRefreshTokenValidity:
-    Type: String
-                      
+    Type: Number
+                  
   userpoolClientWriteAttributes:
     Type: CommaDelimitedList
               
@@ -86,45 +74,21 @@ Parameters:
   userpoolClientSetAttributes:
     Type: String
             
-  useDefault:
+  resourceName:
     Type: String
               
   authSelections:
     Type: String
-              
-  resourceName:
+                                      
+  useDefault:
     Type: String
-                
-  thirdPartyAuth:
-    Type: String
-                    
+                      
   usernameAttributes:
     Type: CommaDelimitedList
-        
-  userPoolGroups:
-    Type: String
               
-  adminQueries:
-    Type: String
-            
-  triggers:
-    Type: String
-                
-  hostedUI:
-    Type: String
-                    
   userPoolGroupList:
     Type: CommaDelimitedList
-                  
-  verificationBucketName:
-    Type: String
-                                            
-  parentStack:
-    Type: String
-                
-  permissions:
-    Type: CommaDelimitedList
-              
+                          
   dependsOn:
     Type: CommaDelimitedList
     
@@ -133,34 +97,13 @@ Conditions:
 
 Resources:
   
-  CustomMessageConfirmationBucket:
-    Type: AWS::S3::Bucket
-    DeletionPolicy: "Retain"
-    Properties:
-      BucketName:  !If [ShouldNotCreateEnvResources, !Ref verificationBucketName, !Join ['',[!Ref verificationBucketName, '-', !Ref env]]]
-      AccessControl: "Private"
-      WebsiteConfiguration:
-        IndexDocument: "index.html"
-        ErrorDocument: "index.html"
-      CorsConfiguration:
-        CorsRules:
-          - 
-            AllowedHeaders:
-              - "Authorization" 
-              - "Content-Length"
-            AllowedMethods:
-              - "GET"
-            AllowedOrigins:
-              - "*"
-            MaxAge: 3000
-  
   
   # BEGIN SNS ROLE RESOURCE
   SNSRole: 
   # Created to allow the UserPool SMS Config to publish via the Simple Notification Service during MFA Process
     Type: AWS::IAM::Role
     Properties:
-      RoleName: !If [ShouldNotCreateEnvResources, 'ZeroCab6dfaf0c_sns-role', !Join ['',['ZeroCab6dfaf0c_sns-role', '-', !Ref env]]]
+      RoleName: !If [ShouldNotCreateEnvResources, 'institbc9982ab_sns-role', !Join ['',['institbc9982ab_sns-role', '-', !Ref env]]]
       AssumeRolePolicyDocument: 
         Version: "2012-10-17"
         Statement: 
@@ -172,10 +115,10 @@ Resources:
               - "sts:AssumeRole"
             Condition: 
               StringEquals:
-                sts:ExternalId: ZeroCab6dfaf0c_role_external_id
+                sts:ExternalId: institbc9982ab_role_external_id
       Policies: 
         - 
-          PolicyName: ZeroCab6dfaf0c-sns-policy
+          PolicyName: institbc9982ab-sns-policy
           PolicyDocument: 
             Version: "2012-10-17"
             Statement: 
@@ -200,25 +143,7 @@ Resources:
           Required: true
           Mutable: true
         
-        -
-          Name: name
-          Required: true
-          Mutable: true
-        
       
-      
-      LambdaConfig:
-        
-        
-          CustomMessage: !Ref functionZeroCarbonCustomMessageArn
-        
-        
-        
-        
-        
-        
-        
-        
       
       
       AutoVerifiedAttributes: !Ref autoVerifiedAttributes
@@ -241,39 +166,15 @@ Resources:
       SmsVerificationMessage: !Ref smsVerificationMessage
       SmsConfiguration: 
         SnsCallerArn: !GetAtt SNSRole.Arn
-        ExternalId: ZeroCab6dfaf0c_role_external_id
+        ExternalId: institbc9982ab_role_external_id
     
-  
-    
-  
-  UserPoolCustomMessageLambdaInvokePermission: 
-    Type: "AWS::Lambda::Permission"
-    DependsOn: UserPool
-    Properties: 
-      Action: "lambda:invokeFunction"
-      Principal: "cognito-idp.amazonaws.com"
-      FunctionName: !Ref functionZeroCarbonCustomMessageName
-      SourceArn: !GetAtt UserPool.Arn
-    
-  
-    
-    
-  
-  
-   
-    
-  # Updating lambda role with permissions to Cognito
-  
   
   UserPoolClientWeb:
   # Created provide application access to user pool
   # Depends on UserPool for ID reference
     Type: "AWS::Cognito::UserPoolClient"
     Properties:
-      ClientName: ZeroCab6dfaf0c_app_clientWeb
-      
-      ReadAttributes: !Ref userpoolClientReadAttributes
-      WriteAttributes: !Ref userpoolClientWriteAttributes
+      ClientName: institbc9982ab_app_clientWeb
       
       RefreshTokenValidity: !Ref userpoolClientRefreshTokenValidity
       UserPoolId: !Ref UserPool
@@ -283,10 +184,7 @@ Resources:
   # Depends on UserPool for ID reference
     Type: "AWS::Cognito::UserPoolClient"
     Properties:
-      ClientName: ZeroCab6dfaf0c_app_client
-      
-      ReadAttributes: !Ref userpoolClientReadAttributes
-      WriteAttributes: !Ref userpoolClientWriteAttributes
+      ClientName: institbc9982ab_app_client
       
       GenerateSecret: !Ref userpoolClientGenerateSecret
       RefreshTokenValidity: !Ref userpoolClientRefreshTokenValidity
@@ -351,7 +249,7 @@ Resources:
   # Marked as depending on UserPoolClientRole for easier to understand CFN sequencing
     Type: 'AWS::IAM::Policy'
     Properties:
-      PolicyName: ZeroCab6dfaf0c_userpoolclient_lambda_iam_policy
+      PolicyName: institbc9982ab_userpoolclient_lambda_iam_policy
       Roles: 
         - !If [ShouldNotCreateEnvResources, !Ref userpoolClientLambdaRole, !Join ['',[!Ref userpoolClientLambdaRole, '-', !Ref env]]]
       PolicyDocument:
@@ -368,7 +266,7 @@ Resources:
   # Marked as depending on UserPoolClientLambdaPolicy for easier to understand CFN sequencing
     Type: 'AWS::IAM::Policy'
     Properties:
-      PolicyName: ZeroCab6dfaf0c_userpoolclient_lambda_log_policy
+      PolicyName: institbc9982ab_userpoolclient_lambda_log_policy
       Roles: 
         - !If [ShouldNotCreateEnvResources, !Ref userpoolClientLambdaRole, !Join ['',[!Ref userpoolClientLambdaRole, '-', !Ref env]]]
       PolicyDocument:
@@ -408,7 +306,7 @@ Resources:
   # Always created
     Type: AWS::Cognito::IdentityPool
     Properties: 
-      IdentityPoolName: !If [ShouldNotCreateEnvResources, 'ZeroCarbon', !Join ['',['ZeroCarbon', '__', !Ref env]]]
+      IdentityPoolName: !If [ShouldNotCreateEnvResources, 'institutoterrabc9982ab_identitypool_bc9982ab', !Join ['',['institutoterrabc9982ab_identitypool_bc9982ab', '__', !Ref env]]]
       
       CognitoIdentityProviders:
         - ClientId:  !Ref UserPoolClient
