--- amplify/backend/function/ZeroCarbonCustomMessage/function-parameters.json
@@ -1,14 +1,0 @@
-{
-    "trigger": true,
-    "modules": [
-        "verification-link"
-    ],
-    "parentResource": "ZeroCarbon",
-    "functionName": "ZeroCarbonCustomMessage",
-    "parentStack": "auth",
-    "triggerEnvs": "[{\"key\":\"RESOURCENAME\",\"value\":\"resourceName\"}]",
-    "triggerTemplate": "CustomMessage.json.ejs",
-    "roleName": "ZeroCarbonCustomMessage",
-    "skipEdit": true,
-    "resourceName": "ZeroCarbonCustomMessage"
-}
