--- amplify/backend/function/ZeroCarbonCustomMessage/parameters.json
@@ -1,4 +1,0 @@
-{
-    "modules": "verification-link",
-    "resourceName": "ZeroCarbonCustomMessage"
-}
