--- amplify/backend/function/ZeroCarbonCustomMessage/src/event.json
@@ -1,12 +1,0 @@
-{
-    "region": "us-west-2",
-    "callerContext": {
-    "clientId": "dummyId"
-    },
-    "userName": "testUser",
-    "triggerSource": "CustomMessage_SignUp",
-    "request": {
-    "codeParameter": "0000"
-    },
-    "response": {}
-  }
