--- amplify/backend/function/ZeroCarbonCustomMessage/src/package.json
@@ -1,7 +1,0 @@
-{
-  "name": "ZeroCarbonCustomMessage",
-  "version": "2.0.0",
-  "description": "Lambda function generated by Amplify",
-  "main": "index.js",
-  "license": "Apache-2.0"
-}
