+++ package.json
@@ -11,19 +11,10 @@
     "aws-amplify": "^1.2.4",
     "aws-amplify-react-native": "^2.2.3",
     "expo": "^35.0.0",
-    "expo-font": "~7.0.0",
-    "native-base": "^2.13.8",
     "react": "16.8.3",
     "react-dom": "16.8.3",
     "react-native": "https://github.com/expo/react-native/archive/sdk-35.0.0.tar.gz",
-    "react-native-gesture-handler": "~1.3.0",
-    "react-native-reanimated": "~1.2.0",
-    "react-native-screens": "^1.0.0-alpha.23",
-    "react-native-web": "^0.11.7",
-    "react-navigation": "^4.0.10",
-    "react-navigation-drawer": "^2.3.3",
-    "react-navigation-stack": "^1.10.3",
-    "react-navigation-tabs": "^2.5.6"
+    "react-native-web": "^0.11.7"
   },
   "devDependencies": {
     "babel-preset-expo": "^7.1.0"
