--- src/utilities/Colors.js
@@ -1,16 +1,0 @@
-const Colors = {
-  white: '#fff',
-  black: '#000',
-
-  darkgrey: '#2f2f2f',
-  grey: '#6d7365',
-  lightgrey: '#f6f6f6',
-
-  green: '#658c1c',
-  lightgreen: '#c7e59a',
-  mutedgreen: '#effbdd',
-
-  lightblue: '#82a2f4'
-}
-
-export default Colors
